refactor(api): move GeolocationController to Api namespace
Move the GeolocationController to the Api namespace to maintain consistency with other API controllers. Also clean up unused imports in routes/api.php
