+++ app/Http/Controllers/Api/GeolocationController.php
@@ -1,13 +1,15 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
 use App\Http\Requests\GeolocationRequest;
-use App\Libraries\IpApi;
+use App\Http\Controllers\Controller;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
 use Illuminate\Http\Request;
+use App\Libraries\IpApi;
 use Maxidev\Logger\TailLogger;
+
 
 class GeolocationController extends Controller
 {
